Add comment scraping and community iteration.
This commit adds functionality for scraping comments during the scraping of submissions. Scaping these at the same time reduces the number of queries.
The commit also add iteration over the datastructure defined in cryptoRegistry.py. This allows the user to input a cryptocurrency and timerange, and the program will iterate over all subrettir/keyword pairs related to the cryptocurrency.