Improve random event generation. Extend correlation explorer to work with multiple trials
Changes to random number generator to produce numbers with the same baseline properties of the changepoint and media-event.
Random numbers are generated within the same bounds as enforced by aggration and pruning during the generation of the true event.
Also extendes the correlation explorer script to support conducting multiple trials in one go, increasing effeciency of testing for
different parameters.